Do not fail release if publishing fails
